Initial support for command-line interface
