+++ Cargo.toml
@@ -5,5 +5,9 @@ authors = ["Justin Shrake <justinshrake@gmail.com>"]
 
 [dependencies]
 glium = "*"
-notify = "^2.5.0"
-time = "0.1"
+notify = "*"
+time = "*"
+clap = "*"
+serde = "*" 
+serde_json = "*" 
+serde_macros = "*"
